[Update] add POST method

+++ api/v1/views/states.py
@@ -7,7 +7,8 @@ from models import storage
 from models.state import State
 
 
-@app_views.route("/states", methods=["GET", "POST"], strict_slashes=False)
+@app_views.route("/states", methods=["GET", "POST"],
+                 strict_slashes=False)
 def states():
     """ Route retrieves list of all State objects or creates State """
     if request.method == "GET":
@@ -22,25 +23,36 @@ def states():
             return "Not a JSON", 400
         if data.get("name") is None:
             return "Missing name", 400
-        return jsonify(State(data).to_dict()), 201
+        obj = State(data)
+        obj.save()
+        return jsonify(obj.to_dict()), 201
 
 
-@app_views.route("/states/<state_id>", methods=["GET"], strict_slashes=False)
-def states_id(state_id):
-    """ Route retrieves a State object """
-    state = storage.get(State, state_id)
-    if state is None:
-        abort(404)
-    return jsonify(state.to_dict())
-
-
-@app_views.route("/states/<state_id>", methods=["DELETE"],
+@app_views.route("/states/<state_id>", methods=["GET", "DELETE", "PUT"],
                  strict_slashes=False)
-def states_id_delete(state_id):
-    """ Route deletes a State object """
-    state = storage.get(State, state_id)
-    if state is None:
-        abort(404)
-    storage.delete(state)
-    storage.save()
-    return jsonify({}), 200
+def states_id(state_id):
+    """ Route retrieves a State object, deletes it or update it """
+    if request.method == "GET":
+        state = storage.get(State, state_id)
+        if state is None:
+            abort(404)
+        return jsonify(state.to_dict())
+    elif request.method == "DELETE":
+        state = storage.get(State, state_id)
+        if state is None:
+            abort(404)
+        storage.delete(state)
+        storage.save()
+        return jsonify({}), 200
+    else:
+        data = request.get_json()
+        if data is None:
+            return "Not a JSON", 400
+        obj = storage.get(State, state_id)
+        if obj is None:
+            abort(404)
+        for key, value in data.items():
+            if key != "id" and key != "created_at" and key != "updated_at":
+                setattr(obj, key, value)
+        obj.save()
+        return jsonify(obj.to_dict()), 200
